Move middleware from /src/utils to /src/middleware
